pyproject.toml
# SPDX-License-Identifier: Apache-2.0

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sns-mcp"
version = "0.1.0"
authors = [{name = "glicoz"}]
description = "Read-only MCP server for Stormshield Network Security (SNS) firewalls"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
keywords = ["mcp", "stormshield", "firewall", "network", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Networking :: Firewalls",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "fastmcp>=2.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "uvicorn>=0.29",
    "stormshield.sns.sslclient",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]

[project.scripts]
sns-mcp = "sns_mcp.server:main_sync"

[tool.hatch.build.targets.wheel]
packages = ["src/sns_mcp"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "stormshield.sns.sslclient",
    "fastmcp",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
